requirements.txt
cachetools==4.2.4
certifi==2021.5.30
charc bset-normalizer==2.0.6
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.2
oauthlib==3.1.1
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.16.0
twilio==6.63.2
uritemplate==4.1.1
urllib3==1.26.7
